pyproject.toml
[project]
name = "aident"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.26.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.0",
]
